+++ anonlink/entitymatch.py
@@ -1,5 +1,4 @@
 import logging
-from operator import itemgetter
 
 from anonlink._entitymatcher import ffi, lib
 
@@ -8,7 +7,7 @@ import sys
 from . import bloommatcher as bm
 from . import util
 
-logging.basicConfig(level=logging.WARNING)
+log = logging.getLogger('anonlink.entitymatch')
 
 
 def python_filter_similarity(filters1, filters2):
@@ -34,15 +33,25 @@ def python_filter_similarity(filters1, filters2):
 
 def cffi_filter_similarity_k(filters1, filters2, k, threshold):
     """Accelerated method for determining Bloom Filter similarity.
+
+    Assumes all filters are the same length, being a multiple of 64
+    bits.
+
     """
     length_f1 = len(filters1)
     length_f2 = len(filters2)
 
-    # We assume the length is 1024 bit = 128 Bytes
-    match_one_against_many_dice_1024_k_top = lib.match_one_against_many_dice_1024_k_top
+    if length_f1 == 0:
+        return []
+
+    filter_bits = len(filters1[0][0])
+    assert(filter_bits % 64 == 0, 'Filter length must be a multple of 64 bits.')
+    filter_bytes = filter_bits // 8
+
+    match_one_against_many_dice_k_top = lib.match_one_against_many_dice_k_top
 
     # An array of the *one* filter
-    clist1 = [ffi.new("char[128]", bytes(f[0].tobytes()))
+    clist1 = [ffi.new("char[{}]".format(filter_bytes), bytes(f[0].tobytes()))
               for f in filters1]
 
     if sys.version_info < (3, 0):
@@ -52,10 +61,10 @@ def cffi_filter_similarity_k(filters1, filters2, k, threshold):
             for b in f[0].tobytes():
                 data.append(b)
 
-        carr2 = ffi.new("char[{}]".format(128 * length_f2), ''.join(data))
+        carr2 = ffi.new("char[{}]".format(filter_bytes * length_f2), ''.join(data))
     else:
         # Works in Python 3+
-        carr2 = ffi.new("char[{}]".format(128 * length_f2),
+        carr2 = ffi.new("char[{}]".format(filter_bytes * length_f2),
                         bytes([b for f in filters2 for b in f[0].tobytes()]))
 
     c_popcounts = ffi.new("uint32_t[{}]".format(length_f2), [f[2] for f in filters2])
@@ -67,18 +76,20 @@ def cffi_filter_similarity_k(filters1, filters2, k, threshold):
 
     result = []
     for i, f1 in enumerate(filters1):
-        assert len(clist1[i]) == 128
-        assert len(carr2) % 64 == 0
-        matches = match_one_against_many_dice_1024_k_top(
+        assert len(clist1[i]) == filter_bytes
+        matches = match_one_against_many_dice_k_top(
             clist1[i],
             carr2,
             c_popcounts,
             length_f2,
+            filter_bytes,
             k,
             threshold,
             c_indices,
             c_scores)
 
+        if matches < 0:
+            raise ValueError('Internel error: Bad key length')
         for j in range(matches):
             ind = c_indices[j]
             assert ind < len(filters2)
@@ -119,7 +130,7 @@ def calculate_mapping_greedy(filters1, filters2, threshold=0.95, k=5):
     :returns A mapping dictionary of index in filters1 to index in filters2.
     """
 
-    logging.info('Solving with greedy solver')
+    log.info('Solving with greedy solver')
 
     sparse_matrix = calculate_filter_similarity(filters1, filters2, k, threshold)
     return greedy_solver(sparse_matrix)