+++ tests/test_bloommatcher.py
@@ -1,8 +1,12 @@
 import unittest
+import pytest
 import random
+import os
+from collections import deque
 from bitarray import bitarray
 
 from anonlink import bloommatcher as bm
+from tests import bitarray_utils
 
 __author__ = 'shardy'
 
@@ -70,6 +74,26 @@ class TestBloomMatcher(unittest.TestCase):
 
         self.assertEqual(result, 0.0)
 
+@pytest.mark.parametrize("L", bitarray_utils.key_lengths)
+def test_dicecoeff(L):
+    """
+    Test the Dice coefficient of bitarrays in bas with other
+    bitarrays of bas.  rotations is the number of times we rotate
+    bas to generate pairs to test the Dice coefficient; 10 takes
+    around 10s, 100 around 60s.
+    """
+    rotations = 100 if "INCLUDE_10K" in os.environ else 10;
+    bas = bitarray_utils.bitarrays_of_length(L)
+
+    # We check that the native code and Python versions of dicecoeff
+    # don't ever differ by more than 10^{-6}.
+    eps = 0.000001
+    d = deque(bas)
+    for _ in range(min(rotations, len(bas))):
+        for a, b in zip(bas, d):
+            diff = bm.dicecoeff_pure_python(a, b) - bm.dicecoeff_native(a, b)
+            assert(abs(diff) < eps)
+        d.rotate(1)
 
 if __name__ == '__main__':
     unittest.main()